Simplify the user-facing API.
- Remove the cdk.Stack subclass
- Rename StackComponent to Stack
- Move `remapCloudControlResource` to an options bag for Stack
