+++ examples/alb/index.ts
@@ -4,9 +4,9 @@ import * as elbv2 from 'aws-cdk-lib/aws-elasticloadbalancingv2';
 import * as pulumi from '@pulumi/pulumi';
 import * as pulumicdk from '@pulumi/cdk';
 import { Construct } from 'constructs';
-import { CfnOutput } from 'aws-cdk-lib';
+import { CfnOutput, Stack } from 'aws-cdk-lib';
 
-class AlbStack extends pulumicdk.Stack {
+class AlbStack extends Stack {
     constructor(scope: Construct, id: string) {
         super(scope, id);
 
@@ -42,6 +42,5 @@ class AlbStack extends pulumicdk.Stack {
     }
 }
 
-const stack = pulumicdk.Stack.create('teststack', AlbStack);
-
+const stack = new pulumicdk.Stack('teststack', AlbStack);
 export const url = stack.outputs['url'];
