+++ examples/cron-lambda/index.ts
@@ -2,16 +2,10 @@ import * as fs from 'fs';
 import * as aws_events from 'aws-cdk-lib/aws-events';
 import * as aws_events_targets from 'aws-cdk-lib/aws-events-targets';
 import * as aws_lambda from 'aws-cdk-lib/aws-lambda';
-import { CfnElement, Duration } from 'aws-cdk-lib';
-import * as pulumi from '@pulumi/pulumi';
-import * as aws from '@pulumi/aws';
-import { Stack } from '@pulumi/cdk';
+import { Duration, Stack } from 'aws-cdk-lib';
+import * as pulumicdk from '@pulumi/cdk';
 import { Construct } from 'constructs';
-
-interface target {
-    arn: pulumi.Input<string>;
-    id: string;
-}
+import { remapCloudControlResource } from './adapter';
 
 class LambdaStack extends Stack {
     constructor(scope: Construct, id: string) {
@@ -32,58 +26,6 @@ class LambdaStack extends Stack {
 
         rule.addTarget(new aws_events_targets.LambdaFunction(lambdaFn));
     }
-
-    public remapCloudControlResource(
-        element: CfnElement,
-        logicalId: string,
-        typeName: string,
-        props: any,
-        options: pulumi.ResourceOptions,
-    ): { [key: string]: pulumi.CustomResource } | undefined {
-        switch (typeName) {
-            case 'AWS::Events::Rule':
-                const resources: { [key: string]: pulumi.CustomResource } = {};
-                const rule = new aws.cloudwatch.EventRule(
-                    logicalId,
-                    {
-                        scheduleExpression: props['scheduleExpression'],
-                        isEnabled: props['state'] == 'ENABLED' ? true : props.State === 'DISABLED' ? false : undefined,
-                        description: props.Description,
-                        eventBusName: props['eventBusName'] ?? undefined,
-                        eventPattern: props['eventPattern'] ?? undefined,
-                        roleArn: props['roleArn'] ?? undefined,
-                    },
-                    options,
-                );
-                resources[logicalId] = rule;
-                const targets: target[] = props['targets'] ?? [];
-                for (const t of targets) {
-                    resources[t.id] = new aws.cloudwatch.EventTarget(
-                        t.id,
-                        {
-                            arn: t.arn,
-                            rule: rule.name,
-                        },
-                        options,
-                    );
-                }
-                return resources;
-            case 'AWS::Lambda::Permission':
-                const perm = new aws.lambda.Permission(
-                    logicalId,
-                    {
-                        action: props['action'],
-                        function: props['functionName'],
-                        principal: props['principal'],
-                        sourceArn: props['sourceArn'] ?? undefined,
-                    },
-                    options,
-                );
-                return { [logicalId]: perm };
-        }
-
-        return undefined;
-    }
 }
 
-const stack = Stack.create('teststack', LambdaStack);
+const stack = new pulumicdk.Stack('teststack', LambdaStack, { remapCloudControlResource });
